pipeline/__init__.py
"""Pipeline Package - Orchestration and configuration"""

from dataclasses import dataclass, field
from typing import List, Dict, Optional, Any
import yaml
import pandas as pd

# config.py
@dataclass
class AnalysisConfig:
    """Configuration for analysis pipeline."""
    
    # Data paths
    data_dir: str
    labels_file: str
    output_dir: str = "results"
    
    # Preprocessing
    window_size: float = 300.0  # seconds (5 minutes)
    overlap: float = 60.0  # seconds (1 minute)
    remove_outliers: bool = True
    apply_filter: bool = True
    
    # Feature extraction
    feature_extractors: List[str] = field(default_factory=lambda: ['breathing'])
    
    # Analysis
    outcomes: List[str] = field(default_factory=lambda: ['consciousness', 'recovery', 'survival'])
    statistical_test: str = 'ttest'
    alpha: float = 0.05
    multiple_comparison_correction: str = 'fdr_bh'
    
    # Dimensionality reduction
    perform_pca: bool = True
    perform_tsne: bool = True
    n_components: int = 2
    
    # Models
    classifiers: List[str] = field(default_factory=lambda: ['svm', 'knn', 'random_forest'])
    hyperparameter_tuning: bool = True
    cv_folds: int = 5
    
    # Visualization
    visualization_backend: str = 'plotly'  # or 'matplotlib'
    create_plots: bool = True
    
    # Export
    export_format: str = 'excel'  # or 'csv'
    
    @classmethod
    def from_yaml(cls, filepath: str) -> 'AnalysisConfig':
        """Load configuration from YAML file."""
        pass
    
    def to_yaml(self, filepath: str) -> None:
        """Save configuration to YAML file."""
        pass
    
    def validate(self) -> None:
        """Validate configuration parameters."""
        pass


# pipeline.py
class AnalysisPipeline:
    """
    Main analysis pipeline - orchestrates entire analysis workflow.
    
    Steps:
    1. Load data and labels
    2. Preprocess (clean, window)
    3. Extract features
    4. Aggregate features
    5. Statistical analysis
    6. Dimensionality reduction
    7. Train classifiers
    8. Evaluate models
    9. Generate visualizations
    10. Export results
    """
    
    def __init__(self, config: AnalysisConfig):
        self.config = config
        self._initialize_components()
    
    def _initialize_components(self) -> None:
        """Initialize all pipeline components based on config."""
        pass
    
    def run(self) -> Dict[str, Any]:
        """
        Run complete analysis pipeline.
        
        Returns:
            Dictionary with all results:
            - features
            - statistical_results
            - dimensionality_reduction_results
            - classification_results
            - visualizations
        """
        pass
    
    def load_data(self) -> tuple:
        """Load recordings and labels."""
        pass
    
    def preprocess_data(self, recordings: list) -> list:
        """Clean and window recordings."""
        pass
    
    def extract_features(self, windows: list) -> 'FeatureCollection':
        """Extract features from windows."""
        pass
    
    def aggregate_features(self, block_features: 'FeatureCollection') -> 'FeatureCollection':
        """Aggregate block features to recording level."""
        pass
    
    def perform_statistical_analysis(
        self, features: 'FeatureCollection', labels: 'ClinicalLabels'
    ) -> Dict[str, Any]:
        """Run statistical tests for all outcomes."""
        pass
    
    def perform_dimensionality_reduction(
        self, features: 'FeatureCollection'
    ) -> Dict[str, Any]:
        """Perform PCA and/or t-SNE."""
        pass
    
    def train_and_evaluate_classifiers(
        self, features: 'FeatureCollection', labels: 'ClinicalLabels'
    ) -> Dict[str, Any]:
        """Train classifiers for all outcomes and evaluate."""
        pass
    
    def generate_visualizations(
        self, features, labels, classification_results, dim_reduction_results
    ) -> None:
        """Generate all plots."""
        pass
    
    def export_results(self, results: Dict[str, Any]) -> str:
        """Export all results to Excel/CSV."""
        pass
    
    def save_state(self, filepath: str) -> None:
        """Save pipeline state for resuming."""
        pass
    
    @classmethod
    def load_state(cls, filepath: str) -> 'AnalysisPipeline':
        """Load saved pipeline state."""
        pass


# batch_processor.py
class BatchProcessor:
    """
    Process multiple analysis experiments.
    
    Useful for running the same analysis with different:
    - Feature sets
    - Classifiers
    - Hyperparameters
    - Window sizes
    """
    
    def __init__(self, base_config: AnalysisConfig):
        self.base_config = base_config
        self.experiments = []
    
    def add_experiment(self, name: str, config_overrides: dict) -> None:
        """Add an experiment with config modifications."""
        pass
    
    def run_all(self, n_jobs: int = 1) -> pd.DataFrame:
        """
        Run all experiments.
        
        Args:
            n_jobs: Number of parallel jobs
            
        Returns:
            DataFrame summarizing all experiment results
        """
        pass
    
    def compare_experiments(self) -> pd.DataFrame:
        """Compare results across experiments."""
        pass


# experiment_manager.py
class ExperimentManager:
    """Manage and track multiple experiments."""
    
    def __init__(self, experiments_dir: str = "experiments"):
        self.experiments_dir = experiments_dir
    
    def create_experiment(self, name: str, config: AnalysisConfig) -> str:
        """Create new experiment directory with config."""
        pass
    
    def list_experiments(self) -> List[str]:
        """List all experiments."""
        pass
    
    def load_experiment(self, name: str) -> 'AnalysisPipeline':
        """Load experiment pipeline."""
        pass
    
    def compare_experiments(self, experiment_names: List[str]) -> pd.DataFrame:
        """Compare results from multiple experiments."""
        pass
