requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0

# Data loading
openpyxl>=3.0.0  # For Excel files
scipy>=1.7.0  # For .mat files

# Visualization
plotly>=5.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Dimensionality reduction
umap-learn>=0.5.0  # For UMAP

# Statistical analysis
statsmodels>=0.13.0
pingouin>=0.5.0  # For advanced statistical tests

# Configuration
pyyaml>=5.4.0

# Progress bars
tqdm>=4.62.0

# Utilities
joblib>=1.1.0  # For parallel processing and model persistence

# Optional: Deep learning (if you want to add neural networks later)
# tensorflow>=2.8.0
# torch>=1.10.0
